updated authorization filter, moved some controllers to other packages
