+++ src/main/java/vlad/petrovskyi/internetshop/web/filters/AuthenticationFilter.java
@@ -42,7 +42,7 @@ public class AuthenticationFilter implements Filter {
 
         Long userId = (Long) req.getSession().getAttribute(USER_ID);
         if (userId == null || userService.get(userId) == null) {
-            resp.sendRedirect("/");
+            resp.sendRedirect("/login");
             return;
         }
