+++ src/main/webapp/web.xml
@@ -194,7 +194,7 @@
 
     <servlet>
         <servlet-name>adminMainPage</servlet-name>
-        <servlet-class>vlad.petrovskyi.internetshop.controllers.AdminPageController</servlet-class>
+        <servlet-class>vlad.petrovskyi.internetshop.controllers.user.AdminPageController</servlet-class>
     </servlet>
     <servlet-mapping>
         <servlet-name>adminMainPage</servlet-name>
@@ -203,7 +203,7 @@
 
     <servlet>
         <servlet-name>accessDenied</servlet-name>
-        <servlet-class>vlad.petrovskyi.internetshop.controllers.AccessDeniedController</servlet-class>
+        <servlet-class>vlad.petrovskyi.internetshop.controllers.user.AccessDeniedController</servlet-class>
     </servlet>
     <servlet-mapping>
         <servlet-name>accessDenied</servlet-name>
